Move Application configuration to Application class
